+++ tests/test_skool2sft.py
@@ -4,9 +4,15 @@ import unittest
 from skoolkittest import SkoolKitTestCase
 from skoolkit import skool2sft, VERSION
 
-def mock_run(*args):
-    global run_args
-    run_args = args
+class MockSftWriter:
+    def __init__(self, *args):
+        global mock_sft_writer
+        self.args = args
+        self.write_called = False
+        mock_sft_writer = self
+
+    def write(self):
+        self.write_called = True
 
 class Skool2SftTest(SkoolKitTestCase):
     def test_no_arguments(self):
@@ -21,12 +27,14 @@ class Skool2SftTest(SkoolKitTestCase):
             self.assertTrue(error.startswith('usage: skool2sft.py'))
 
     def test_default_option_values(self):
-        self.mock(skool2sft, 'run', mock_run)
+        self.mock(skool2sft, 'SftWriter', MockSftWriter)
         skoolfile = 'test.skool'
         skool2sft.main((skoolfile,))
-        infile, write_hex = run_args
+        infile, write_hex, preserve_base = mock_sft_writer.args
         self.assertEqual(infile, skoolfile)
         self.assertFalse(write_hex)
+        self.assertFalse(preserve_base)
+        self.assertTrue(mock_sft_writer.write_called)
 
     def test_option_V(self):
         for option in ('-V', '--version'):
@@ -36,13 +44,26 @@ class Skool2SftTest(SkoolKitTestCase):
             self.assertEqual(error[0], 'SkoolKit {}'.format(VERSION))
 
     def test_option_h(self):
-        self.mock(skool2sft, 'run', mock_run)
+        self.mock(skool2sft, 'SftWriter', MockSftWriter)
         skoolfile = 'test.skool'
         for option in ('-h', '--hex'):
             skool2sft.main((option, skoolfile))
-            infile, write_hex = run_args
+            infile, write_hex, preserve_base = mock_sft_writer.args
             self.assertEqual(infile, skoolfile)
             self.assertTrue(write_hex)
+            self.assertFalse(preserve_base)
+            self.assertTrue(mock_sft_writer.write_called)
+
+    def test_option_b(self):
+        self.mock(skool2sft, 'SftWriter', MockSftWriter)
+        skoolfile = 'test.skool'
+        for option in ('-b', '--preserve-base'):
+            skool2sft.main((option, skoolfile))
+            infile, write_hex, preserve_base = mock_sft_writer.args
+            self.assertEqual(infile, skoolfile)
+            self.assertFalse(write_hex)
+            self.assertTrue(preserve_base)
+            self.assertTrue(mock_sft_writer.write_called)
 
     def test_run(self):
         skool = '; Routine\nc32768 RET'